+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * :bug: Fix missing stacktraces when throwing bare strings ([#324](https://github.com/Roblox/jest-roblox/pull/324))
+* :bug: Fix bad error message when test contains syntax error ([#326](https://github.com/Roblox/jest-roblox/pull/326))
 
 ## 3.2.2 (2022-01-04)
 * :bug: Fix debugging capabilities broken recently ([#325](https://github.com/Roblox/jest-roblox/pull/325))